@@ -1,25 +1,54 @@
-#include <stdarg.h>
-#include <stdio.h>
-#include <stdlib.h>
+#include <cstdarg>
+#include <cstdio>
+#include <cstdlib>
 
+#include "logging.h"
 #include "modules.h"
 
-#include <h/amul.alog.h>
+// Support breakpoints instead of exits for errors
+#ifndef __has_builtin
+#define __has_builtin(x) 0
+#endif
 
-extern void Terminate(error_t err);
+#ifndef NDEBUG
+#if !defined(DEBUG_BREAK) && defined(_MSC_VER)
+#define DEBUG_BREAK __debugbreak()
+#endif
+#if !defined(DEBUG_BREAK) && !defined(_MSC_VER)
+#if __has_builtin(__builtin_debugtrap)
+#define DEBUG_BREAK __builtin_debugtrap()
+#endif
+#if !defined(DEBUG_BREAK) && defined(POSIX)
+#include <signal.h>
+#if defined(SIGTRAP)
+#define DEBUG_BREAK raise(SIGTRAP)
+#else
+#define DEBUG_BREAK raise(SIGABRT)
+#endif
+#endif
+#endif
+#endif
+
+#if !defined(DEBUG_BREAK)
+#define DEBUG_BREAK (void) 0
+#endif
+
+[[noreturn]] void Terminate(error_t err);
 
 // Track the number of messages logged with AL_ERROR
-uint32_t al_errorCount = 0;
+static size_t s_errorCount{ 0 };
 
-static const char *levelName[MAX_LOG_LEVEL + 1] = {"Debug", "Info",  "Note",    "WARNING",
-                                                   "ERROR", "FATAL", "DISABLED"};
+static LogLevel s_logLevel = LogLevel::LWARN;
 
-static LogLevel s_logLevel = MAX_LOG_LEVEL;
+constexpr string_view levelName[MAX_LOG_LEVEL + 1] = {
+    // Names of the debugging levels
+    "More", "Debug", "Info", "Note", "WARNING", "ERROR", "FATAL", "DISABLED"
+};
 
 error_t
 loggingModuleInit(Module * /*module*/)
 {
-    s_logLevel = AL_INFO;
+    s_logLevel = LNOTE;
     return 0;
 }
 
@@ -32,7 +61,7 @@ loggingModuleStart(Module * /*module*/)
         return EINVAL;
     }
 
-    alog(AL_DEBUG, "Logging enabled");
+    LogDebug("Logging enabled");
 
     return 0;
 }
@@ -40,7 +69,7 @@ loggingModuleStart(Module * /*module*/)
 error_t
 loggingModuleClose(Module * /*module*/, error_t /*err*/)
 {
-    alog(AL_DEBUG, "Logging disabled");
+    LogDebug("Logging disabled");
     s_logLevel = MAX_LOG_LEVEL;
     return 0;
 }
@@ -48,52 +77,64 @@ loggingModuleClose(Module * /*module*/, error_t /*err*/)
 error_t
 InitLogging()
 {
-    return NewModule(
-            MOD_LOGGING, loggingModuleInit, loggingModuleStart, loggingModuleClose, NULL, NULL);
+    return NewModule(MOD_LOGGING,
+                     loggingModuleInit,
+                     loggingModuleStart,
+                     loggingModuleClose,
+                     nullptr,
+                     nullptr);
 }
 
 void
-alogLevel(LogLevel level)
+LogBreak()
+{
+    DEBUG_BREAK;
+}
+
+size_t
+GetLogErrorCount() noexcept
+{
+    return s_errorCount;
+}
+
+void
+BumpLogErrorCount() noexcept
+{
+    ++s_errorCount;
+}
+
+LogLevel
+GetLogLevel() noexcept
+{
+    return s_logLevel;
+}
+
+void
+SetLogLevel(LogLevel level) noexcept
 {
     if (level < MAX_LOG_LEVEL)
         s_logLevel = level;
 }
 
-const char *
-alogGetLevelName()
+string_view
+GetLogLevelName(LogLevel level) noexcept
 {
-    return levelName[s_logLevel];
+    return levelName[level];
 }
 
 void
-alog(LogLevel level, const char *fmt, ...)
+LogFinish(LogLevel level)
 {
-    if (level < MAX_LOG_LEVEL && level >= s_logLevel) {
-        if (level != AL_INFO)
-            printf("%c- ", levelName[level][0]);
-        else
-            printf("-- ");
-
-        va_list vl;
-        va_start(vl, fmt);
-        vprintf(fmt, vl);
-        va_end(vl);
-        printf("\n");
-        fflush(stdout);
-
-        switch (level) {
-        case AL_ERROR:
-            al_errorCount++;
-            break;
-        case AL_FATAL:
-            al_errorCount++;
-            break;
-        default:
-            break;
-        }
+    if (level >= LERROR) {
+        std::cout << std::endl;
+        ++s_errorCount;
+    } else {
+        std::cout << "\n";
     }
-    if (level >= AL_FATAL) {
-        DEBUG_BREAK;
-        Terminate(-1);
+    if (level >= LFATAL) {
+        std::cout << std::flush;
+        fflush(stdout);
+        LogBreak();
+        exit(1);
     }
 }
