@@ -1,15 +1,14 @@
-#include <assert.h>
-#include <stdio.h>
-#include <stdlib.h>
+#include <cassert>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
 
+#include "amigastubs.h"
+#include "amul.test.h"
+#include "typedefs.h"
+#include "logging.h"
 #include "modules.h"
 #include "system.h"
-
-#include <h/amul.alog.h>
-#include <h/amul.test.h>
-#include <h/amul.type.h>
-
-#include <h/amigastubs.h>
 
 Module *s_modulesHead;
 Module *s_modulesTail;
@@ -18,11 +17,11 @@ bool s_modulesInitialized;
 bool s_modulesClosed;
 
 static const char *moduleNames[MAX_MODULE_ID] = {
-        // hard-coded names for modules
-        "INVALID", "logging", "cmdline", "strings", "compiler",
+    // hard-coded names for modules
+    "INVALID", "logging", "cmdline", "strings", "compiler", "server", "client",
 };
 
-void
+[[noreturn]] void
 Terminate(error_t err)
 {
     CloseModules(err);
@@ -40,13 +39,13 @@ InitModules()
 error_t
 StartModules()
 {
-    for (Module *cur = s_modulesHead; cur; cur = (Module *)cur->links.next) {
+    for (Module *cur = s_modulesHead; cur; cur = (Module *) cur->links.next) {
         if (cur->start) {
-            alog(AL_DEBUG, "Starting Module #%d: %s", cur->id, cur->name);
+            LogDebug("Starting Module #", cur->id, ": ", cur->name);
             error_t err = cur->start(cur);
             if (err != 0) {
-                alog(AL_ERROR, "Module initialization failed: aborting.");
-                DEBUG_BREAK;
+                LogError("Module initialization failed: aborting.");
+                LogBreak();
                 Terminate(err);
             }
         }
@@ -59,12 +58,13 @@ CloseModules(error_t err)
 {
     Module *prev = nullptr;
     for (Module *cur = s_modulesTail; cur; cur = prev) {
-        alog(AL_DEBUG, "Closing Module #%d: %s", cur->id, cur->name);
-        prev = (Module *)cur->links.prev;
+        LogDebug("Closing Module #", cur->id, ": ", cur->name);
+        prev = (Module *) cur->links.prev;
         error_t reterr = CloseModule(cur, err);
         if (reterr != 0) {
-            fprintf(stderr, "*** INTERNAL ERROR: Module %s failed to terminate with %d\n",
-                    cur->name, reterr);
+            // NOTE: Use
+            std::cerr << "*** INTERNAL ERROR: Module " << cur->name
+                      << " failed to terminate: " << reterr << std::endl;
         }
     }
 
@@ -72,28 +72,31 @@ CloseModules(error_t err)
 }
 
 error_t
-NewModule(
-        ModuleID id, moduleinit_fn init /*opt*/, modulestart_fn start /*opt*/,
-        moduleclose_fn close /*opt*/, void *context /*opt*/, Module **ptr /*opt*/)
+NewModule(ModuleID id,
+          moduleinit_fn init /*opt*/,
+          modulestart_fn start /*opt*/,
+          moduleclose_fn close /*opt*/,
+          void *context /*opt*/,
+          Module **ptr /*opt*/)
 {
     REQUIRE(id && id < MAX_MODULE_ID);
     REQUIRE(context || (init || start || close));
 
-    if (GetModule(id) != NULL) {
-        alog(AL_DEBUG, "Tried to register duplicate module#%d: %s", id, moduleNames[id]);
+    if (GetModule(id) != nullptr) {
+        LogDebug("Tried to register duplicate module#", id, ": ", id, moduleNames[id]);
         return EEXIST;
     }
 
-    Module *cur = (Module *)AllocateMem(sizeof(Module));
+    Module *cur = (Module *) AllocateMem(sizeof(Module));
     if (!cur) {
-        afatal("Out of memory");
+        LogFatal("Out of memory");
     }
 
     // populate values
     cur->id = id;
     cur->name = moduleNames[id];
-    cur->links.next = NULL;
-    cur->links.prev = (DoubleLinkedNode *)s_modulesTail;
+    cur->links.next = nullptr;
+    cur->links.prev = (DoubleLinkedNode *) s_modulesTail;
     cur->init = init;
     cur->start = start;
     cur->close = close;
@@ -103,15 +106,15 @@ NewModule(
         s_modulesHead = cur;
         s_modulesTail = cur;
     } else {
-        s_modulesTail->links.next = (DoubleLinkedNode *)cur;
-        cur->links.prev = (DoubleLinkedNode *)s_modulesTail;
+        s_modulesTail->links.next = (DoubleLinkedNode *) cur;
+        cur->links.prev = (DoubleLinkedNode *) s_modulesTail;
         s_modulesTail = cur;
     }
 
     if (cur->init) {
         error_t err = cur->init(cur);
         if (err != 0) {
-            afatal("Module #%d: %s: initialization failed: %d", id, cur->name, err);
+            LogFatal("Module #", id, ": ", cur->name, ": initialization failed: ", err);
         }
     }
     if (ptr)
@@ -123,11 +126,11 @@ NewModule(
 Module *
 GetModule(ModuleID id)
 {
-    for (Module *cur = s_modulesHead; cur; cur = (Module *)cur->links.next) {
+    for (Module *cur = s_modulesHead; cur; cur = (Module *) cur->links.next) {
         if (id == cur->id)
             return cur;
     }
-    return NULL;
+    return nullptr;
 }
 
 error_t
@@ -138,7 +141,7 @@ CloseModule(Module *module, error_t err)
     // Make sure this is a registered module
     Module *cur = s_modulesHead;
     while (cur && cur != module)
-        cur = (Module *)cur->links.next;
+        cur = (Module *) cur->links.next;
     if (cur != module)
         return EFAULT;
 
@@ -151,13 +154,13 @@ CloseModule(Module *module, error_t err)
     if (module->links.next)
         module->links.next->prev = module->links.prev;
     if (s_modulesHead == module)
-        s_modulesHead = (Module *)module->links.next;
+        s_modulesHead = (Module *) module->links.next;
     if (s_modulesTail == module)
-        s_modulesTail = (Module *)module->links.prev;
+        s_modulesTail = (Module *) module->links.prev;
 
     memset(module, 0, sizeof(*module));
 
-    ReleaseMem((void **)&module);
+    ReleaseMem((void **) &module);
 
     return reterr;
 }
@@ -165,5 +168,5 @@ CloseModule(Module *module, error_t err)
 error_t
 RegisterContextModule(ModuleID id, void *context)
 {
-    return NewModule(id, NULL, NULL, NULL, context, NULL);
+    return NewModule(id, nullptr, nullptr, nullptr, context, nullptr);
 }
