@@ -1,45 +1,21 @@
-#include <stdio.h>
-#include <sys/stat.h>
-#if defined(_MSC_VER)
-#    include <direct.h>
-#else
-#    include <unistd.h>
-#endif
+#include <cstdio>
+#include <cstring>
 
-#include "amulcom.h"
+#include "amul.argp.h"
+#include "amul.file.h"
+#include "filesystem.h"
+#include "logging.h"
 #include "modules.h"
 #include "system.h"
-
-#include <h/amul.alog.h>
-#include <h/amul.argp.h>
-
-// For systems that support it: CTRL-C handler.
-void
-CXBRK()
-{
-    fprintf(stderr, "*** CTRL-C pressed: terminating\n");
-    Terminate(EINTR);
-}
-
-error_t
-misuse(const char **argv, const char *issue, const char *arg, error_t err)
-{
-    fprintf(stderr, "%s: %s: %s\n", argv[0], issue, arg);
-    fprintf(stderr, "Try '%s -help' for usage information\n", argv[0]);
-    fflush(stderr);
-    return err;
-}
 
 error_t
 usage(const char **argv)
 {
-    printf("Usage: %s [-h|-?] [-d] [-v|-q] [-r] [game directory]\n", argv[0]);
+    printf("Usage: %s [-h|-?] [-v|-q] [game directory]\n", argv[0]);
     printf("Compiler for AMUL multi-player games.\n");
     printf("\n");
     printf("  -help, -?                 Displays this help information\n");
-    printf("  -d, -dmoves               Enables 'dmove' checking\n");
     printf("  -q, -quiet                Decreases output verbosity\n");
-    printf("  -r, -room-reuse           Re-uses previous room data during compilation\n");
     printf("  -v, -verbose              Increases output verbosity\n");
     printf("  <game directory>          Optional path game files and data will be in\n");
 
@@ -49,74 +25,53 @@ usage(const char **argv)
 error_t
 ParseCommandLine(const CommandLine *cmdline)
 {
-    const int    argc = cmdline->argc;
+    const int argc = cmdline->argc;
     const char **argv = cmdline->argv;
-    struct stat  sb;
-
-    int desiredLogLevel = AL_INFO;
+    struct stat sb;
 
     for (int n = 1; n < argc; n++) {
         const char *arg = argv[n];
         if (arg[0] == '-') {
             if (strncmp("-h", arg, 2) == 0 || strncmp("--h", arg, 2) == 0 || strcmp("-?", arg) == 0)
                 return usage(argv);
-
-            if (strncmp("-d", arg, 2) == 0) {
-                checkDmoves = true;
-                continue;
-            }
             if (strncmp("-q", arg, 2) == 0) {
-                if (desiredLogLevel < MAX_LOG_LEVEL - 1)
-                    ++desiredLogLevel;
+                if (g_desiredLogLevel < MAX_LOG_LEVEL - 1)
+                    ++g_desiredLogLevel;
                 continue;
             }
             if (strncmp("-v", arg, 2) == 0) {
-                if (desiredLogLevel > 0)
-                    --desiredLogLevel;
+                if (g_desiredLogLevel > 0)
+                    --g_desiredLogLevel;
                 continue;
             }
-            if (strcmp("-r", arg) == 0) {
-                reuseRoomData = true;
-                continue;
-            }
-            return misuse(argv, "Unrecognized argument", arg, EINVAL);
+            return CmdlineMisuse(argv, "Unrecognized argument", arg, EINVAL);
         }
-        if (gameDir[0] != 0)
-            return misuse(argv, "Unexpected argument", arg, EINVAL);
-        if (PathCopy(gameDir, sizeof(gameDir), NULL, arg) != 0)
-            return misuse(argv, "Invalid game directory", arg, EINVAL);
-        int err = stat(gameDir, &sb);
-        if (err != 0)
-            return misuse(argv, "Missing game directory/no access", gameDir, err);
+        if (!gameDir.empty())
+            return CmdlineMisuse(argv, "Unexpected argument", arg, EINVAL);
+        PathAdd(gameDir, arg);
+        if (auto err = stat(gameDir.c_str(), &sb); err != 0)
+            return CmdlineMisuse(argv, "Missing game directory/no access", gameDir.c_str(), err);
         if (S_ISREG(sb.st_mode))
-            return misuse(argv, "Not a directory", gameDir, EINVAL);
+            return CmdlineMisuse(argv, "Not a directory", gameDir.c_str(), EINVAL);
     }
 
     char pwd[MAX_PATH_LENGTH];
     if (!getcwd(pwd, sizeof(pwd)))
-        afatal("Cannot get CWD");
+        LogFatal("Cannot get CWD");
 
-    if (strncmp(gameDir, "./", 2) == 0) {
-        path_concater(pwd, gameDir + 2);
-        gameDir[0] = 0;
+    if (gameDir.substr(0, 2) == "./") {
+        PathJoin(gameDir, pwd, gameDir.substr(2));
     }
-    if (gameDir[0] == 0 || strcmp(gameDir, ".") == 0) {
-        path_copier(gameDir, pwd);
+    if (gameDir.empty()) {
+#if defined(AMUL_DEFAULT_GAME)
+        gameDir = AMUL_DEFAULT_GAME;
+#else
+        gameDir = ".";
+#endif
     }
-
-    alogLevel((LogLevel)desiredLogLevel);
+    if (gameDir == ".") {
+        gameDir = pwd;
+    }
 
     return 0;
-}
-
-error_t
-initCommandLine(Module *module)
-{
-    return ParseCommandLine((CommandLine *)(module->context));
-}
-
-error_t
-InitCommandLine(const CommandLine *cmdline)
-{
-    return NewModule(MOD_CMDLINE, initCommandLine, NULL, NULL, (void *)cmdline, NULL);
 }
