@@ -1,4 +1,4 @@
-#include <h/amul.xtra.h>
+#include "amul.xtra.h"
 
 ///////////////////////////////////////////////////////////////////////////////
 // converted from extras.i
@@ -11,7 +11,7 @@ extractLine(const char *from, char *to)
     for (;;) {
         if (!isCommentChar(*from))
             break;
-        from = strstop((char *)from, '\n');
+        from = strstop((char *) from, '\n');
     }
     // copy the text into to
     while (*from && !isEol(*from)) {
