@@ -1,15 +1,16 @@
-#ifndef AMUL_SRC_AMULCOM_FILEPROCESSING_H
-#define AMUL_SRC_AMULCOM_FILEPROCESSING_H
+#ifndef AMULCOM_FILEPROCESSING_H
+#define AMULCOM_FILEPROCESSING_H
 
 #include <cstdio>
-extern FILE *ifp;
 
-extern char Word[64];    ///TODO: Pass
+extern thread_local FILE *ifp;
+
+extern char Word[64];  /// TODO: Pass
 
 bool nextc(bool required);
-char *getTidyBlock(FILE *fp);
+char *getTidiedLineToScratch(FILE *fp);
 void tidy(char *ptr);
-void skipblock();
-void checkErrorCount();
+void skipParagraph();
+void CheckErrorCount();
 
-#endif  // AMUL_SRC_AMULCOM_FILEPROCESSING_H
+#endif  // AMULCOM_FILEPROCESSING_H
