@@ -1,14 +1,14 @@
-#include "amulcom.h"
+#include "amul.argp.h"
+#include "amul.test.h"
+#include "logging.h"
 #include "modules.h"
 
-#include <h/amul.alog.h>
-#include <h/amul.argp.h>
-#include <h/amul.test.h>
+extern int amulcom_main();
 
 int
 main(int argc, const char **argv)
 {
-    CommandLine cmdline = {argc, argv, nullptr};
+    CommandLine cmdline = { argc, argv, nullptr };
 
     InitModules();
 
