@@ -1,30 +1,22 @@
-#ifndef AMUL_SRC_AMULCOM_STRINGS_H
-#define AMUL_SRC_AMULCOM_STRINGS_H 1
+#pragma once
+#ifndef AMULCOM_STRINGS_H
+#define AMULCOM_STRINGS_H 1
 
-#include <h/amul.strs.h>
-#include <h/amul.type.h>
+#include "amul.strs.h"
+#include "typedefs.h"
 
-#include <cstdbool>
-#include <cstdio>
+error_t InitStrings() noexcept;
 
-error_t InitStrings();
+size_t GetStringCount() noexcept;
+size_t GetStringBytes() noexcept;
 
-size_t GetStringCount();
-size_t GetStringBytes();
+error_t AddTextString(string_view text, bool isLine, stringid_t *idp);
 
-error_t AddTextString(const char *start, const char *end, bool isLine, stringid_t *idp);
+error_t
+RegisterTextString(string_view label, string_view text, bool isLine, stringid_t *idp);
 
-error_t RegisterTextString(
-        const char *label, const char *start, const char *end, bool isLine, stringid_t *idp);
-static inline
-error_t RegisterTextString(
-        const char *label, const char *start, size_t size, bool isLine, stringid_t *idp)
-{
-    return RegisterTextString(label, start, start + size, isLine, idp);
-}
+error_t TextStringFromFile(const char *label, FILE *fp, stringid_t *idp, bool toEol = false);
 
-error_t TextStringFromFile(const char *label, FILE *fp, stringid_t *idp, bool toEol=false);
-
-error_t LookupTextString(const char *label, stringid_t *idp=nullptr);
+error_t LookupTextString(string_view label, stringid_t *idp = nullptr);
 
 #endif
