@@ -1,29 +1,24 @@
-#ifndef AMUL_SRC_BUFFER_H
-#define AMUL_SRC_BUFFER_H
-
+#ifndef AMUL_BUFFER_H
+#define AMUL_BUFFER_H
 // Buffer is for consuming bytes from a fixed range in memory.
 
-#include <h/amul.type.h>
+#include <cstring>
+
+#include "typedefs.h"
 
 struct Buffer {
-    const char *m_start{nullptr};
-    const char *m_cur{nullptr};
-    const char *m_end{nullptr};
+    const char *m_start{ nullptr };
+    const char *m_cur{ nullptr };
+    const char *m_end{ nullptr };
 
     constexpr Buffer() noexcept {}
     constexpr Buffer(const char *start, const char *end) noexcept
-        : m_start{start}
-        , m_cur{start}
-        , m_end{end}
+        : m_start{ start }, m_cur{ start }, m_end{ end }
     {
     }
-    constexpr Buffer(const char *start, size_t len) noexcept
-        : Buffer{start, start + len}
-    {
-    }
-    template <size_t Size>
-    constexpr Buffer(const char (&str)[Size]) noexcept
-        : Buffer{&str[0], &str[0] + Size}
+    constexpr Buffer(const char *start, size_t len) noexcept : Buffer{ start, start + len } {}
+    template<size_t Size>
+    constexpr Buffer(const char (&str)[Size]) noexcept : Buffer{ &str[0], &str[0] + Size }
     {
     }
 
@@ -38,14 +33,30 @@ struct Buffer {
     constexpr const char *end() const noexcept { return m_end; }
     constexpr const char *it() const noexcept { return m_cur; }
 
-    constexpr bool   Eof() const noexcept { return it() >= end(); }
+    constexpr bool Eof() const noexcept { return it() >= end(); }
     constexpr size_t Size() const noexcept { return end() - begin(); }
 
     uint8_t Peek() const noexcept { return !Eof() ? *m_cur : 0; }
     uint8_t Read() noexcept { return (!Eof()) ? *(m_cur++) : 0; }
-    void    Skip() noexcept { m_cur += !Eof() ? 1 : 0; }
+    void Skip() noexcept { m_cur += !Eof() ? 1 : 0; }
+
+    const char *ReadLine() noexcept
+    {
+        if (Eof())
+            return nullptr;
+        const char *eol = strpbrk(m_cur, "\r\n");
+        if (eol) {
+            m_cur = eol;
+            if (*m_cur == '\r')
+                ++m_cur;
+            if (*m_cur == '\n')
+                ++m_cur;
+        } else
+            eol = m_end;
+        return eol;
+    }
 
     void Close() noexcept { Assign(nullptr, 0); }
 };
 
-#endif  // AMUL_SRC_BUFFER_H
+#endif  // AMUL_BUFFER_H
