@@ -5,37 +5,37 @@
 
 // Generated by gen_charmap.c
 constexpr AtomType charToAtom[256] = {
-        A_END,     A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_SPACE,   A_END,     A_INVALID, A_INVALID, A_END,     A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_SPACE,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
-        A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
-        A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,
-        A_DIGIT,   A_DIGIT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
-        A_PUNCT,   A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
-        A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
-        A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
-        A_LETTER,  A_LETTER,  A_LETTER,  A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
-        A_PUNCT,   A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
-        A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
-        A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
-        A_LETTER,  A_LETTER,  A_LETTER,  A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
-        A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_END,     A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_SPACE,   A_END,     A_INVALID, A_INVALID, A_END,     A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_SPACE,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
+    A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
+    A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,   A_DIGIT,
+    A_DIGIT,   A_DIGIT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
+    A_PUNCT,   A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
+    A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
+    A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
+    A_LETTER,  A_LETTER,  A_LETTER,  A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,
+    A_PUNCT,   A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
+    A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
+    A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,  A_LETTER,
+    A_LETTER,  A_LETTER,  A_LETTER,  A_PUNCT,   A_PUNCT,   A_PUNCT,   A_PUNCT,   A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
+    A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID, A_INVALID,
 };
 #endif  // CHAR_ATOM_MAP_H
