@@ -1,7 +1,8 @@
-#ifndef AMUL_SRC_TOKENSTREAM_H
-#define AMUL_SRC_TOKENSTREAM_H
+#ifndef AMUL_TOKENSTREAM_H
+#define AMUL_TOKENSTREAM_H
 
-#include "array.h"
+#include <vector>
+
 #include "atom.h"
 #include "buffer.h"
 #include "token.h"
@@ -10,17 +11,20 @@ struct TokenStream {
     Buffer &buffer;
 
     // The atoms we have available
-    Array<Atom, 256>      atoms{};
-    Array<TokenType, 256> tokens{};
-    Array<TokenType, 64>  newTokens{};
+    std::vector<Atom> atoms{};
+    std::vector<TokenType> tokens{};
+    std::vector<TokenType> newTokens{};
 
     TokenStream(Buffer &_buffer)
-        : buffer(_buffer)
+        : buffer(_buffer) 
     {
+        atoms.reserve(256);
+        tokens.reserve(256);
+        newTokens.reserve(64);
     }
 
-    void *m_pattern{nullptr};
-    void  SetPattern(void *pattern) { m_pattern = pattern; }
+    void *m_pattern{ nullptr };
+    void SetPattern(void *pattern) { m_pattern = pattern; }
 
   protected:
     void Atomize();  // Transform next line into atoms
@@ -28,4 +32,4 @@ struct TokenStream {
     void ScanLine();  // Transform next line into tokens
 };
 
-#endif  // AMUL_SRC_TOKENSTREAM_H
+#endif  // AMUL_TOKENSTREAM_H
